+++ src/pmp/io/read_off.h
@@ -1,0 +1,14 @@
+// Copyright 2011-2022 the Polygon Mesh Processing Library developers.
+// Distributed under a MIT-style license, see LICENSE.txt for details.
+
+#pragma once
+
+#include <string>
+
+#include "pmp/SurfaceMesh.h"
+
+namespace pmp {
+
+void read_off(SurfaceMesh& mesh, const std::string& filename);
+
+} // namespace pmp