+++ CHANGELOG.md
@@ -10,6 +10,10 @@ This project aims to adhere to [Semantic Versioning](https://semver.org/spec/v2.
 
 ### Changed
 
+- Remove support for PLY format, see 1ec912b for why.
+- Remove support for XYZ and AGI point set formats.
+- Move `SurfaceMesh::read()` and `SurfaceMesh::write()` to free functions `pmp::read()` and `pmp::write()`.
+- Move IO functionality into its own module
 - Upgrade C++ standard to C++17
 - Remove `get_*_property_type()` functions from `SurfaceMesh`
 - Move example apps from `src/apps` to `examples`. Remove `PMP_BUILD_APPS` CMake option.